--- src/main/java/Task.java
@@ -1,11 +1,0 @@
-public abstract class Task {
-    public boolean isMarked() {
-        return isMarked;
-    }
-    public void setMarked(boolean marked) {
-        isMarked = marked;
-    }
-     abstract void initializeMessage();
-    abstract void displayMessage();
-    boolean isMarked;
-}
